+++ pyklip/klip.py
@@ -428,7 +428,7 @@ def rotate(img, angle, center, new_center=None, flipx=False, astr_hdr=None):
 
     #create the coordinate system of the image to manipulate for the transform
     dims = img.shape
-    x, y = np.meshgrid(np.arange(dims[1], dtype=np.float32), np.arange(dims[0], dtype=np.float32))
+    x, y = np.meshgrid(np.arange(dims[1], dtype=float), np.arange(dims[0], dtype=float))
 
     #if necessary, move coordinates to new center
     if new_center is not None:
